update BA codes and fix IPC new process
Add ElectricityLCI v2 to dependencies and utilize the read_ba_codes method (better than the CSV file in this repo). New find_new_process_id method to query the openLCA database or JSON-LD (tested on IPC server; not on JSON-LD) after a process is added---the UUID is not auto generated until after the netl.add method is called for IPC connections. Added new section to make_outlook_gen to check that database/JSON-LD write/update worked and that the new process ID is found.